also needs EnsureReadable in DecodeSymbolAddress!
